.Trash-1000/files/plc-temperature-reader/requirements.txt
pymcprotocol
python-dotenv
redis
prometheus_client